Avoid dynamic imports
